@@ -13,77 +13,67 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-material-components/file-input": "^16.0.1",
-    "@angular/animations": "17.1.1",
-    "@angular/cdk": "17.1.1",
-    "@angular/common": "17.1.1",
-    "@angular/compiler": "17.1.1",
-    "@angular/core": "17.1.1",
-    "@angular/fire": "17.0.1",
-    "@angular/forms": "17.1.1",
-    "@angular/localize": "17.1.1",
-    "@angular/material": "17.1.1",
-    "@angular/platform-browser": "17.1.1",
-    "@angular/platform-browser-dynamic": "17.1.1",
-    "@angular/router": "17.1.1",
-    "@fortawesome/fontawesome-free": "6.6.0",
-    "@fortawesome/fontawesome-svg-core": "6.6.0",
-    "@fortawesome/free-brands-svg-icons": "6.6.0",
-    "@fortawesome/free-regular-svg-icons": "6.6.0",
-    "@fortawesome/free-solid-svg-icons": "6.6.0",
-    "angular-calendar": "0.31.1",
-    "animate.css": "4.1.1",
-    "date-fns": "4.1.0",
+    "@angular/animations": "^18.2.12",
+    "@angular/cdk": "^18.2.13",
+    "@angular/common": "^18.2.12",
+    "@angular/compiler": "^18.2.12",
+    "@angular/core": "^18.2.12",
+    "@angular/fire": "^18.0.1",
+    "@angular/forms": "^18.2.12",
+    "@angular/localize": "^18.2.12",
+    "@angular/material": "^18.2.13",
+    "@angular/platform-browser": "^18.2.12",
+    "@angular/platform-browser-dynamic": "^18.2.12",
+    "@angular/router": "^18.2.12",
+    "@fortawesome/fontawesome-free": "^6.6.0",
+    "@fortawesome/fontawesome-svg-core": "^6.6.0",
+    "@fortawesome/free-brands-svg-icons": "^6.6.0",
+    "@fortawesome/free-regular-svg-icons": "^6.6.0",
+    "@fortawesome/free-solid-svg-icons": "^6.6.0",
+    "angular-calendar": "^0.31.1",
+    "animate.css": "^4.1.1",
+    "audio4live": "file:",
+    "date-fns": "^4.1.0",
     "email-scramble": "^3.0.0",
-    "firebase": "10.14.1",
-    "mdb-angular-ui-kit": "6.1.0",
+    "firebase": "^11.0.2",
+    "mdb-angular-ui-kit": "^7.1.0",
     "mjml": "^4.15.3",
     "mjml-browser": "^4.15.3",
-    "ng-recaptcha": "13.2.1",
-    "ngx-material-timepicker": "13.1.1",
-    "ngx-quicklink": "0.4.5",
-    "ngx-quill": "24.0.5",
-    "quill": "1.3.7",
-    "rxjs": "7.8.1",
-    "tslib": "2.8.1",
-    "zone.js": "0.14.6"
+    "ng-recaptcha-2": "^14.0.0",
+    "ngx-material-timepicker": "^13.1.1",
+    "ngx-quicklink": "^0.4.6",
+    "ngx-quill": "^26.0.10",
+    "quill": "^2.0.2",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.8.1",
+    "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "0.1800.1",
-    "@angular-devkit/build-angular": "17.3.10",
-    "@angular-eslint/builder": "17.5.1",
-    "@angular-eslint/eslint-plugin": "17.5.1",
-    "@angular-eslint/eslint-plugin-template": "17.5.1",
-    "@angular-eslint/schematics": "17.5.1",
-    "@angular-eslint/template-parser": "17.5.1",
-    "@angular/cli": "17.3.8",
-    "@angular/compiler-cli": "17.1.1",
-    "@angular/language-service": "18.0.0",
-    "@types/node": "22.9.0",
-    "@types/quill": "2.0.14",
-    "@typescript-eslint/eslint-plugin": "8.14.0",
-    "@typescript-eslint/parser": "8.14.0",
-    "eslint": "^8.57.0",
-    "firebase-tools": "13.25.0",
-    "husky": "9.1.6",
-    "inquirer-autocomplete-prompt": "3.0.1",
-    "lint-staged": "15.2.10",
-    "modernizr": "3.13.1",
-    "prettier": "3.3.3",
-    "prettier-plugin-organize-imports": "4.1.0",
-    "ts-node": "10.9.2",
-    "typescript": "5.3.3",
+    "@angular-devkit/architect": "^0.1802.12",
+    "@angular-eslint/builder": "^18.4.0",
+    "@angular-eslint/eslint-plugin": "^18.4.0",
+    "@angular-eslint/eslint-plugin-template": "^18.4.0",
+    "@angular-eslint/schematics": "^18.4.0",
+    "@angular-eslint/template-parser": "^18.4.0",
+    "@angular/build": "^18.2.12",
+    "@angular/cli": "^18.2.12",
+    "@angular/compiler-cli": "^18.2.12",
+    "@angular/language-service": "^18.2.12",
+    "@types/node": "^22.9.0",
+    "@types/quill": "^2.0.14",
+    "@typescript-eslint/eslint-plugin": "^8.15.0",
+    "@typescript-eslint/parser": "^8.15.0",
+    "eslint": "^9.15.0",
+    "firebase-tools": "^13.25.0",
+    "husky": "^9.1.7",
+    "inquirer-autocomplete-prompt": "^3.0.1",
+    "lint-staged": "^15.2.10",
+    "modernizr": "^3.13.1",
+    "prettier": "^3.3.3",
+    "prettier-plugin-organize-imports": "^4.1.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.5.4",
     "webpack-bundle-analyzer": "^4.10.2"
-  },
-  "overrides": {
-    "@angular-material-components/file-input": {
-      "@angular/platform-browser": ">=16.0.0",
-      "@angular/common": ">=16.0.0",
-      "@angular/core": ">=16.0.0",
-      "@angular/forms": ">=16.0.0",
-      "@angular/material": ">=16.0.0",
-      "@angular/cdk": ">=16.0.0"
-    }
   },
   "husky": {
     "hooks": {
