@@ -10,7 +10,7 @@ import { adapterFactory } from 'angular-calendar/date-adapters/date-fns';
 import { MdbCheckboxModule } from 'mdb-angular-ui-kit/checkbox';
 import { MdbFormsModule } from 'mdb-angular-ui-kit/forms';
 import { MdbModalModule } from 'mdb-angular-ui-kit/modal';
-import { RECAPTCHA_LANGUAGE, RecaptchaFormsModule, RecaptchaModule } from 'ng-recaptcha';
+import { RECAPTCHA_LANGUAGE, RecaptchaFormsModule, RecaptchaModule } from 'ng-recaptcha-2';
 import { NgxMaterialTimepickerModule, TIME_LOCALE } from 'ngx-material-timepicker';
 import { environment } from '../../environments/environment';
 import { BuchenRoutingModule } from './buchen-routing.module';
@@ -23,7 +23,7 @@ registerLocaleData(localeDe);
   imports: [
     CommonModule,
     BuchenRoutingModule,
-    provideFirebaseApp(() => initializeApp(environment.firebase)),
+    FormsModule,
     FormsModule,
     RecaptchaModule,
     RecaptchaFormsModule,
@@ -33,6 +33,17 @@ registerLocaleData(localeDe);
     MdbCheckboxModule,
     FirestoreModule,
     FunctionsModule,
+    CalendarModule.forRoot({ provide: DateAdapter, useFactory: adapterFactory })
+  ],
+  providers: [
+    {
+      provide: RECAPTCHA_LANGUAGE,
+      useValue: 'de'
+    },
+    {
+      provide: TIME_LOCALE,
+      useValue: 'de-DE'
+    },
     provideFunctions(() => {
       const functions = getFunctions();
       if (environment.useEmulators) {
@@ -47,17 +58,7 @@ registerLocaleData(localeDe);
       }
       return firestore;
     }),
-    CalendarModule.forRoot({ provide: DateAdapter, useFactory: adapterFactory })
-  ],
-  providers: [
-    {
-      provide: RECAPTCHA_LANGUAGE,
-      useValue: 'de'
-    },
-    {
-      provide: TIME_LOCALE,
-      useValue: 'de-DE'
-    }
+    provideFirebaseApp(() => initializeApp(environment.firebase))
   ]
 })
 export class BuchenModule {}
